quick typo fix

+++ authnzerver/actions/user.py
@@ -625,6 +625,7 @@ def delete_user(payload,
         users.c.user_id,
         users.c.email,
         users.c.password,
+        users.c.user_role
     ]).select_from(
         users
     ).where(
@@ -661,6 +662,16 @@ def delete_user(payload,
             'messages':["We could not verify your email address or password."]
         }
 
+    if row['user_role'] == 'superuser':
+        LOGGER.error(
+            "Can't delete superusers."
+        )
+        return {
+            'success': False,
+            'user_id':payload['user_id'],
+            'email':payload['email'],
+            'messages':["Can't delete superusers."]
+        }
 
     # delete the user
     delete = users.delete().where(
@@ -675,9 +686,7 @@ def delete_user(payload,
 
     # don't forget to delete the sessions as well
     delete = sessions.delete().where(
-        users.c.user_id == payload['user_id']
-    ).where(
-        users.c.user_role != 'superuser'
+        sessions.c.user_id == payload['user_id']
     )
     result = currproc.connection.execute(delete)
     result.close()
